+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -3,7 +3,7 @@ package kitchenpos.application;
 import java.util.List;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.ui.dto.menu.MenuGroupCreateRequest;
+import kitchenpos.application.dto.menu.MenuGroupCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
