+++ src/main/java/kitchenpos/application/dto/order/OrderStatusChangeRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.order;
+package kitchenpos.application.dto.order;
 
 public class OrderStatusChangeRequest {
