+++ src/main/java/kitchenpos/application/dto/ordertable/OrderTableChangeNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.ordertable;
+package kitchenpos.application.dto.ordertable;
 
 public class OrderTableChangeNumberOfGuestsRequest {
