+++ src/main/java/kitchenpos/application/dto/ordertable/OrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.ordertable;
+package kitchenpos.application.dto.ordertable;
 
 public class OrderTableCreateRequest {
