+++ src/main/java/kitchenpos/application/dto/ordertable/OrderTableIdDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.ordertable;
+package kitchenpos.application.dto.ordertable;
 
 public class OrderTableIdDto {
