+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -4,8 +4,8 @@ import java.net.URI;
 import java.util.List;
 import kitchenpos.application.OrderService;
 import kitchenpos.domain.Order;
-import kitchenpos.ui.dto.order.OrderCreateRequest;
-import kitchenpos.ui.dto.order.OrderStatusChangeRequest;
+import kitchenpos.application.dto.order.OrderCreateRequest;
+import kitchenpos.application.dto.order.OrderStatusChangeRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
