+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import java.net.URI;
 import kitchenpos.application.TableGroupService;
 import kitchenpos.domain.TableGroup;
-import kitchenpos.ui.dto.tablegroup.TableGroupCreateRequest;
-import kitchenpos.ui.dto.tablegroup.TableGroupUngroupRequest;
+import kitchenpos.application.dto.tablegroup.TableGroupCreateRequest;
+import kitchenpos.application.dto.tablegroup.TableGroupUngroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
