+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -4,9 +4,9 @@ import java.net.URI;
 import java.util.List;
 import kitchenpos.application.TableService;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.ui.dto.ordertable.OrderTableChangeEmptyRequest;
-import kitchenpos.ui.dto.ordertable.OrderTableChangeNumberOfGuestsRequest;
-import kitchenpos.ui.dto.ordertable.OrderTableCreateRequest;
+import kitchenpos.application.dto.ordertable.OrderTableChangeEmptyRequest;
+import kitchenpos.application.dto.ordertable.OrderTableChangeNumberOfGuestsRequest;
+import kitchenpos.application.dto.ordertable.OrderTableCreateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
