+++ src/test/java/kitchenpos/support/fixture/dto/TableGroupCreateRequestFixture.java
@@ -3,8 +3,8 @@ package kitchenpos.support.fixture.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.ui.dto.ordertable.OrderTableIdDto;
-import kitchenpos.ui.dto.tablegroup.TableGroupCreateRequest;
+import kitchenpos.application.dto.ordertable.OrderTableIdDto;
+import kitchenpos.application.dto.tablegroup.TableGroupCreateRequest;
 
 public class TableGroupCreateRequestFixture {
